Fix pulse tail; enable dynamic lights on all ship thrusters
Pulse weapon visual effects were previously being displayed with the tail leading and the head at the end. This update corrects that so that pulses now display as a leading head and a trailing tail.

This update also enables the dynamic thruster lighting effect on all ships, as the performance impact turns out to be low compared to the cost of enabling 4x supersampling.