Overlay Mode interaction updates
The game now has multiple display modes, which need to interact nicely with each other. There's the basic Flight Mode, but now we can also turn on the overlay mode (with the icons at the top), and that has to work correctly with pressing Space to pause the game, and with Esc to bring up the game dialog (which also pauses the game), and when the ship is Docked.