Fix NewGame options being overwritten on first display by UI
The slider options in the Settings.lua views -- Audio, Graphics, Interface -- correctly use as an initial slider value the value in GameState.X.Y for each variable. However, the slider options in NewGame.lua would not do this -- they are being read in correctly from user.ini, but the value initially assigned to the slider for each New Game option was whatever its Config.X.Y default was.

After 3 frustrating hours, I finally narrowed down the cause: the sliders displayed in Settings.lua are called from within a function, and the sliders displayed in NewGame.lua are not. It looks like there's some kind of early assignment that happens when the View function addContent() is first called that uses the default variable values _before_ they get overwritten by the values from user.ini... but when you wrap the Container UI code in a function, the initial slider values are set only _after_ InitFiles.lua reads the user's preferred values from user.ini.

The latter is the behavior we want. So I have moved the slider UI generation code in NewGame.lua into a function.

*However*, I remember your point that we'd really rather not do this -- it's better to let the UI code be run "early" and not wrap it in a function (with all the overhead that creates).

The New Game settings are broken without the change I've made here. But we should come back to this later with a solution that lets us avoid wrapping the UI code in a function just to force variable initialization after reading user value preferences from user.ini.

Extra note: the little change to Slider.lua to insure that self.state.lastValue has been initialized before testing it resolves a separate problem I noticed, which is that every slider's callback function was being called over and over again as soon as the slider was created. When lastValue doesn't exist yet because the user hasn't clicked on the slider for the first time yet, the test always fails and the slider's callback gets needlessly executed. Checking for the existence of lastValue before testing it resolves this issue.