Move to new cargo build syntax, and add the link search path for libphx.
Rust 1.81 seemed to change the linker paths, so was failing to link
against the phx library. Potentally a change to the working directory.
To fix this, this commit explicitly adds the linker path to the phx
library using cargo::rustc-link-search.

At the same time, as our MSV is >1.77, we can move to the _new_ Cargo
build syntax: cargo::<key>